Create ScopeDescriptor to isolate scope content and delegate its usage
ScopeDescriptor is an abstraction to identify what are the roots and children of the scope. By having this abstraction and ScopeGraph its implementation, we can avoid manual delegation.
