+++ core/src/main/kotlin/motif/core/ScopeGraph.kt
@@ -15,7 +15,6 @@
  */
 package motif.core
 
-import motif.ast.IrClass
 import motif.ast.IrType
 import motif.models.ChildMethod
 import motif.models.ErrorScope
@@ -25,10 +24,27 @@ import motif.models.Scope
 class ScopeEdge(val parent: Scope, val child: Scope, val method: ChildMethod)
 
 /**
+ * A representation of Scope, what is consist of and what it contains
+ */
+interface ScopeDescriptor {
+
+    val roots: List<Scope>
+
+    val scopes: List<Scope>
+
+    fun getScope(scopeType: IrType): Scope?
+
+    fun getChildEdges(scope: Scope): Iterable<ScopeEdge>
+
+    fun getParentEdges(scope: Scope): Iterable<ScopeEdge>
+
+}
+
+/**
  * Graph of [Scopes][Scope] as defined by [Scope.childMethods]. Throws an [IllegalStateException] if any of a
  * Scope's childEdge Scopes does not exist in the initial list of Scopes.
  */
-internal class ScopeGraph private constructor(val scopes: List<Scope>) {
+internal class ScopeGraph private constructor(override val scopes: List<Scope>) : ScopeDescriptor {
 
     private val scopeMap: Map<IrType, Scope> = scopes.associateBy { it.clazz.type }
     private val childEdges: Map<Scope, List<ScopeEdge>> = scopes.associate { scope -> scope to createChildren(scope) }
@@ -37,21 +53,21 @@ internal class ScopeGraph private constructor(val scopes: List<Scope>) {
         scopes.associateWith { mapping[it] ?: emptyList() }
     }()
 
-    val roots: List<Scope> = parentEdges.filter { it.value.isEmpty() }.map { it.key }
+    override val roots: List<Scope> = parentEdges.filter { it.value.isEmpty() }.map { it.key }
 
     val scopeCycleError: ScopeCycleError? = calculateCycle()
 
     val parsingErrors: List<ParsingError> = scopes.filterIsInstance<ErrorScope>().map { it.parsingError }
 
-    fun getChildEdges(scope: Scope): List<ScopeEdge> {
+    override fun getChildEdges(scope: Scope): List<ScopeEdge> {
         return childEdges[scope] ?: throw NullPointerException("Scope not found: ${scope.qualifiedName}")
     }
 
-    fun getParentEdges(scope: Scope): List<ScopeEdge> {
+    override fun getParentEdges(scope: Scope): List<ScopeEdge> {
         return parentEdges[scope] ?: throw NullPointerException("Scope not found: ${scope.qualifiedName}")
     }
 
-    fun getScope(scopeType: IrType): Scope? {
+    override fun getScope(scopeType: IrType): Scope? {
         return scopeMap[scopeType]
     }
